Update ap-classes.service.ts
I added difficulty to the ApClass Interface and I added difficulty into the ApClass array. This should eventually go on to store other metrics like the score distribution as I referenced in the ap-insights-api.md. The code for that isn't that difficult, but it should be hard coded and you shouldn't do anything like web scraping as it is too expensive of an operation for the backend to handle on a regular basis.