Implemented Back-end of the User Inputting Already Taken AP Classes
The feature works by having a dedicated endpoint in your ap-classes.controller.ts file that listens for POST requests at /ap-classes/taken. Users send a list of AP classes they've taken to this endpoint. The server receives this list, and for now, logs it to the console and returns a success status, indicating the data was received.